File size and downloaded size should be reported as float, not as integer

+++ DownloadManager/Plugin.DownloadManager.Android/DownloadManagerImplementation.cs
@@ -160,8 +160,8 @@ namespace Plugin.DownloadManager
          */
         public void UpdateFileProperties (ICursor cursor, DownloadFileImplementation downloadFile)
         {
-            downloadFile.TotalBytesWritten = cursor.GetInt (cursor.GetColumnIndex (Android.App.DownloadManager.ColumnBytesDownloadedSoFar));
-            downloadFile.TotalBytesExpected = cursor.GetInt (cursor.GetColumnIndex (Android.App.DownloadManager.ColumnTotalSizeBytes));
+            downloadFile.TotalBytesWritten = cursor.GetFloat (cursor.GetColumnIndex (Android.App.DownloadManager.ColumnBytesDownloadedSoFar));
+            downloadFile.TotalBytesExpected = cursor.GetFloat (cursor.GetColumnIndex (Android.App.DownloadManager.ColumnTotalSizeBytes));
 
             switch ((DownloadStatus)cursor.GetInt (cursor.GetColumnIndex (Android.App.DownloadManager.ColumnStatus))) {
             case DownloadStatus.Successful:
